refactor: [#581] move semantic validation
to the type that has the knowledge.
